Safari-609.1.11.1/LayoutTests/http/tests/media/media-stream/device-change-event-in-iframe-expected.txt

     
<iframe id=camera> got a 'devicechange' event
<iframe id=main> got a 'devicechange' event
<iframe id=microphone> got a 'devicechange' event
<iframe id=same-origin> got a 'devicechange' event

PASS: 'devicechange' only fired on frames with access to devices
